Weekly schedule view now displays meetings properly
accepts meeting objects now and can display them based on their start dates.  it can't handle meetings that overlap between days, and it is restricting to 15 minute increments. also it can't handle overlapping meetings, so we need to think of a way to fix that (or not, maybe its not important?)